Move #3 - files related to auth
인증과 관련된 파일들을 적절한 패키지에 이동
