+++ src/main/java/me/kevinntech/onlineshop/auth/CurrentUserArgumentResolver.java
@@ -1,5 +1,7 @@
 package me.kevinntech.onlineshop.auth;
 
+import me.kevinntech.onlineshop.auth.annotation.CurrentUser;
+import me.kevinntech.onlineshop.auth.dto.LoginUser;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
