+++ src/main/java/me/kevinntech/onlineshop/auth/annotation/UniqueNickname.java
@@ -1,4 +1,6 @@
-package me.kevinntech.onlineshop.auth;
+package me.kevinntech.onlineshop.auth.annotation;
+
+import me.kevinntech.onlineshop.auth.validator.UniqueNicknameValidator;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
