+++ src/main/java/me/kevinntech/onlineshop/auth/dto/LoginUser.java
@@ -1,4 +1,4 @@
-package me.kevinntech.onlineshop.auth;
+package me.kevinntech.onlineshop.auth.dto;
 
 import lombok.Builder;
 import lombok.Getter;
